create base class ServiceAudience
